requirements.txt
scikit-learn
ftfy
regex
tqdm
huggingface-hub
transformers
numpy<2.0
pandas
open-clip-torch
torch
torchvision
webdataset
img2dataset
datasets
jupyter
ipywidgets
matplotlib
seaborn
wadllib
Pillow
tensorboard
kaggle
fsspec==2023.9.2
pyyaml